+++ app/models/criterion.rb
@@ -1,0 +1,11 @@
+class Criterion < ApplicationRecord
+  # Add prefix ID according to conventions
+  has_prefix_id :crit
+
+  # Associations
+  belongs_to :rubric
+  has_many :levels, dependent: :destroy
+
+  # Validations
+  validates :title, presence: true
+end