+++ changelog.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2025-04-20]
+### Added
+- Implement `Criterion` model with attributes (`title`, `description`, `position`), associations (`rubric`, `levels`), validation (`title`), and prefixed ID (`crit_`).
+- Add migration for `criteria` table, enforcing `NOT NULL` on `title` and `rubric_id`.
+- Add model tests and fixtures for `Criterion`.
+- Add `criterion`/`criteria` inflection rule.
+- Update `fixtures.mdc` rule to prevent invalid fixtures.
+
 ## [2026-04-19]
 ### Added
 - Create `bin/rails check` Rake task to run tests, Rubocop, and Brakeman.